requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
websockets==12.0
httpx==0.27.2
duckduckgo-search==6.2.6
python-dotenv==1.0.1
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
markdown==3.6
Jinja2==3.1.4
weasyprint==62.3
aiofiles==24.1.0
typing-extensions==4.12.2
bleach==6.1.0
Pygments==2.18.0